+++ lib/src/database/twonly.db.dart
@@ -82,7 +82,7 @@ class TwonlyDB extends _$TwonlyDB {
   TwonlyDB.forTesting(DatabaseConnection super.connection);
 
   @override
-  int get schemaVersion => 18;
+  int get schemaVersion => 19;
 
   static QueryExecutor _openConnection() {
     final connection = driftDatabase(
@@ -237,6 +237,12 @@ class TwonlyDB extends _$TwonlyDB {
             await m.addColumn(
               schema.contacts,
               schema.contacts.askForFriendPromotions,
+            );
+          },
+          from18To19: (m, schema) async {
+            await m.addColumn(
+              schema.keyVerifications,
+              schema.keyVerifications.verifiedBy,
             );
           },
         )(m, from, to);